backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
transformers>=4.35.0
torch>=2.0.0
pycryptodome>=3.19.0
hypothesis>=6.92.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pydantic>=2.0.0
python-multipart>=0.0.6
